DEV-189 #time 30m separating main into peripheral specific sub-files
